move MCP approval prompts back to MCP
