+++ codex-rs/core/src/codex_delegate_tests.rs
@@ -1,6 +1,6 @@
 use super::*;
-use crate::guardian::MCP_TOOL_APPROVAL_DECLINE_SYNTHETIC;
-use crate::guardian::MCP_TOOL_APPROVAL_QUESTION_ID_PREFIX;
+use crate::mcp_tool_call::MCP_TOOL_APPROVAL_DECLINE_SYNTHETIC;
+use crate::mcp_tool_call::MCP_TOOL_APPROVAL_QUESTION_ID_PREFIX;
 use async_channel::bounded;
 use codex_protocol::config_types::ApprovalsReviewer;
 use codex_protocol::models::NetworkPermissions;
