+++ codex-rs/core/src/guardian/mod.rs
@@ -26,50 +26,7 @@ use serde::Serialize;
 pub(crate) use approval_request::GuardianApprovalRequest;
 pub(crate) use approval_request::GuardianMcpAnnotations;
 pub(crate) use approval_request::GuardianNetworkAccessTrigger;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_ACCEPT;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_ACCEPT_AND_REMEMBER;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_ACCEPT_FOR_SESSION;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_CANCEL;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_CONNECTOR_DESCRIPTION_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_CONNECTOR_ID_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_CONNECTOR_NAME_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_DECLINE_SYNTHETIC;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_KIND_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_KIND_MCP_TOOL_CALL;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_PERSIST_ALWAYS;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_PERSIST_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_PERSIST_SESSION;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_QUESTION_ID_PREFIX;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_SOURCE_CONNECTOR;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_SOURCE_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_TOOL_DESCRIPTION_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_TOOL_PARAMS_DISPLAY_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_TOOL_PARAMS_KEY;
-#[allow(unused_imports)]
-pub(crate) use approval_request::MCP_TOOL_APPROVAL_TOOL_TITLE_KEY;
-pub(crate) use approval_request::McpToolApprovalPromptOptions;
 pub(crate) use approval_request::guardian_approval_request_to_json;
-pub(crate) use approval_request::is_mcp_tool_approval_question_id;
-pub(crate) use approval_request::mcp_tool_approval_question_id;
 pub(crate) use review::guardian_rejection_message;
 pub(crate) use review::guardian_timeout_message;
 pub(crate) use review::is_guardian_reviewer_source;
